+++ CHANGELOG.md
@@ -7,7 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-Nothing yet.
+### Added
+
+- Added new appearance, backup, folder, globus, info, moon, open link, and storage icons
 
 ## [0.10.1+11] - 2026-07-20